work around ironmq returning status=deleted messages
Things appear to be broken and the normal "get" calls are returning
already deleted messages.  There's no way to get rid of them short of
clearing the queue.  This is causing us lots of grief and IronMQ has
made it clear they're not going to fix it (insisting we switch all our
services to their v3 service instead).

Work around this by adding Status to the Message struct and skipping
over deleted messages when returning to the caller of the Get() variant
methods.  The deleted items will be re-added to the queue when the
timeout expires, but there's nothing that can be done about that because
deleting a deleted message has no effect.  In the meantime we can make
progress on other messages.